// src/app/languages/interface/more/orderDetails.interface.ts
export interface IOrderDetails {
    orderDetails: string;
    returnOrder: string;
    clickHere: string;
    enterreasonForReturn: string;
    submit: string;
    returnRemarks: string;
    orderDate: string;
    orderNo: string;
    orderTotal: string;
    items: string;
    downloadInvoice: string;
    shippingDetails: string;
    trackOrder: string;
    paymentInformation: string;
    paymentType: string;
    billingAddress: string;
    shippingAddress: string;
    orderSummary: string;
    itemsCost: string;
    deliveryCharges: string;
    tax: string;  
}